@@ -1,9 +1,10 @@
 package com.jasik.momsnaggingapi.domain.diary;
 
-import com.jasik.momsnaggingapi.domain.common.BaseTime;
+import com.jasik.momsnaggingapi.infra.common.BaseTime;
 import com.jasik.momsnaggingapi.domain.diary.Diary.DailyResponse;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDate;
+import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -18,28 +19,19 @@ import org.springframework.format.annotation.DateTimeFormat;
 @Entity
 @NamedNativeQuery(name = "findDailyResponse", query =
     "select IF(a.title is null and a.context is null, false, true) as diaryExists, b.diaryDate\n"
-        + "from\n"
-        + "    (select *\n"
-        + "     from diary\n"
-        + "     where user_id = :userId\n"
-        + "       and diary_date >= :startDate\n"
-        + "       and diary_date <= :endDate\n"
-        + "     ) a\n"
-        + "        right outer join (\n"
-        + "        SELECT\n"
+        + "from\n" + "    (select *\n" + "     from diary\n" + "     where user_id = :userId\n"
+        + "       and diary_date >= :startDate\n" + "       and diary_date <= :endDate\n"
+        + "     ) a\n" + "        right outer join (\n" + "        SELECT\n"
         + "            DATE_FORMAT(DATE_ADD(:startDate, INTERVAL seq - 1 DAY), '%Y-%m-%d') AS diaryDate\n"
         + "        FROM (SELECT @num \\:= @num + 1 AS seq\n"
         + "              FROM information_schema.tables a\n"
         + "                 , information_schema.tables b\n"
-        + "                 , (SELECT @num \\:= 0) c\n"
-        + "             ) T\n"
+        + "                 , (SELECT @num \\:= 0) c\n" + "             ) T\n"
         + "        WHERE seq <=  DATEDIFF(:endDate, :startDate) + 1\n"
-        + "    ) b on a.diary_date = b.diaryDate;",
-    resultSetMapping = "DailyResponse")
+        + "    ) b on a.diary_date = b.diaryDate;", resultSetMapping = "DailyResponse")
 @SqlResultSetMapping(name = "DailyResponse", classes = @ConstructorResult(targetClass = DailyResponse.class, columns = {
     @ColumnResult(name = "diaryExists", type = Boolean.class),
-    @ColumnResult(name = "diaryDate", type = LocalDate.class)
-}))
+    @ColumnResult(name = "diaryDate", type = LocalDate.class)}))
 @Getter
 @Setter
 @NoArgsConstructor
@@ -53,6 +45,7 @@ public class Diary extends BaseTime {
     private Long userId;
 
     private String title;
+    @Column(columnDefinition = "TEXT")
     private String context;
     private LocalDate diaryDate;
 
@@ -64,12 +57,12 @@ public class Diary extends BaseTime {
         this.diaryDate = diaryDate;
     }
 
-    public void updateDiary(String title, String context){
+    public void updateDiary(String title, String context) {
         this.title = title;
         this.context = context;
     }
 
-    public void initUser(Long userId){
+    public void initUser(Long userId) {
         this.userId = userId;
     }
 
@@ -79,15 +72,14 @@ public class Diary extends BaseTime {
     @AllArgsConstructor
     @NoArgsConstructor
     public static class DiaryRequest {
-
         @Schema(description = "일기장 제목", defaultValue = "오랜만의 일기")
-        private String title;
-
-        @Schema(description = "일기장 내용", defaultValue = "오랜만에 일기를 써본다. 재밌었다.")
-        private String context;
-
-        @Schema(description = "일기장 날짜")
         @NotNull
+        private String title;
+        @Schema(description = "일기장 내용", defaultValue = "오랜만에 일기를 써본다. 재밌었다.")
+        @NotNull
+        private String context;
+        @Schema(description = "일기장 날짜")
+        @NotBlank
         @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
         private LocalDate diaryDate;
     }
@@ -106,7 +98,7 @@ public class Diary extends BaseTime {
         private String context;
 
         @Schema(description = "일기장 날짜")
-        @NotNull
+        @NotBlank
         @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
         private LocalDate diaryDate;
 
@@ -114,6 +106,7 @@ public class Diary extends BaseTime {
             "false"})
         private boolean today;
     }
+
     @Schema(description = "특정 월의 일기장 작성 여부 조회 시 응답 클래스")
     @Getter
     @Setter
