@@ -1,28 +1,50 @@
 package com.jasik.momsnaggingapi.domain.schedule;
 
-import com.jasik.momsnaggingapi.domain.common.BaseTime;
+import com.jasik.momsnaggingapi.domain.schedule.Schedule.SchedulePush;
+import com.jasik.momsnaggingapi.infra.common.BaseTime;
+import com.jasik.momsnaggingapi.infra.common.ErrorCode;
+import com.jasik.momsnaggingapi.infra.common.exception.NotValidRoutineException;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import javax.persistence.Column;
+import javax.persistence.ColumnResult;
+import javax.persistence.ConstructorResult;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.validation.constraints.NotNull;
+import javax.persistence.SqlResultSetMapping;
+import javax.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.GenericGenerator;
+import org.hibernate.annotations.NamedNativeQuery;
 import org.hibernate.annotations.Parameter;
 import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.format.annotation.DateTimeFormat.ISO;
+import org.springframework.lang.Nullable;
 
 
 @Entity
+@NamedNativeQuery(
+    name = "findSchedulePush",
+    query = "select a.schedule_name as title, "
+        + "       b.firebase_token as targetToken, "
+        + "       case b.nagging_level when 1 then c.level1 when 2 then c.level2 when 3 then c.level3 else c.level1 end as body "
+        + "from schedule a inner join user b on a.user_id = b.id inner join nagging c on a.nagging_id = c.id\n"
+        + "where a.schedule_date = :scheduleDate "
+        + "and a.alarm_time = :alarmTime",
+    resultSetMapping = "SchedulePush")
+@SqlResultSetMapping(name = "SchedulePush", classes = @ConstructorResult(targetClass = SchedulePush.class, columns = {
+    @ColumnResult(name = "title", type = String.class),
+    @ColumnResult(name = "targetToken", type = String.class),
+    @ColumnResult(name = "body", type = String.class)}))
 @Getter
 @Setter
 @NoArgsConstructor
@@ -44,8 +66,7 @@ public class Schedule extends BaseTime {
 
     private Long categoryId;
 
-    @Column(columnDefinition = "int default 0")
-    private int seqNumber;
+    private Long naggingId;
 
     @Column(columnDefinition = "int default 0")
     private int goalCount;
@@ -58,7 +79,7 @@ public class Schedule extends BaseTime {
 
     private String scheduleTime;
 
-    @Column(nullable = false)
+//    @Column(nullable = false)
     private LocalDate scheduleDate;
 
     private LocalTime alarmTime;
@@ -86,16 +107,42 @@ public class Schedule extends BaseTime {
     @Enumerated(EnumType.STRING)
     private ScheduleType scheduleType;
 
+    public void initOriginalId() {
+        this.originalId = this.id;
+    }
+
+    public void initScheduleDate(LocalDate nextDate) {
+        this.scheduleDate = nextDate;
+    }
+    public void initScheduleTypeAndUserId(Long userId){
+        if (this.goalCount > 0 | this.mon | this.tue | this.wed | this.thu | this.fri | this.sat | this.sun){
+            this.scheduleType = ScheduleType.ROUTINE;
+        }else{
+            this.scheduleType = ScheduleType.TODO;
+        }
+        this.userId = userId;
+    }
+    public void initGoalCount() {
+        this.goalCount = 0;
+    }
+    // TODO: getter 안먹히는 컬럼들 -> _ 와 연관이 있는 듯
+    public boolean getDone() {
+        return this.done;
+    }
+    public boolean plusDoneCount() {
+        this.doneCount += 1;
+        return this.doneCount >= this.goalCount;
+    }
+
     @Builder
-    public Schedule(Long userId, Long originalId, Long categoryId, int seqNumber, int goalCount,
+    public Schedule(Long userId, Long originalId, Long categoryId, int goalCount,
         int doneCount, String scheduleName, String scheduleTime, LocalDate scheduleDate,
         LocalTime alarmTime, boolean done,
-//                    LocalDateTime routineEndDate,
-        boolean mon, boolean tue, boolean wed, boolean thu, boolean fri, boolean sat, boolean sun) {
+        boolean mon, boolean tue, boolean wed, boolean thu, boolean fri, boolean sat, boolean sun,
+        long naggingId) {
         this.userId = userId;
         this.originalId = originalId;
         this.categoryId = categoryId;
-        this.seqNumber = seqNumber;
         this.scheduleName = scheduleName;   // 수정가능 -> 이후 전부 변경 -> update all -> index 유지
         this.scheduleTime = scheduleTime;   // 수정가능 -> 이후 전부 변경 -> update all -> index 유지
         this.scheduleDate = scheduleDate;   // 수정가능 -> 타겟만 변경 -> update
@@ -110,34 +157,13 @@ public class Schedule extends BaseTime {
         this.fri = fri;   // 수정가능 -> 이후 전부 변경 -> delete -> create
         this.sat = sat;   // 수정가능 -> 이후 전부 변경 -> delete -> create
         this.sun = sun;   // 수정가능 -> 이후 전부 변경 -> delete -> create
-    }
-
-    public void initOriginalId() {
-        this.originalId = this.id;
-    }
-
-    public void initScheduleDate(LocalDate nextDate) {
-        this.scheduleDate = nextDate;
-    }
-
-    public boolean plusDoneCount() {
-        this.doneCount += 1;
-        return this.doneCount >= this.goalCount;
+        this.naggingId = naggingId;
     }
 
     public void initNextSchedule() {
-        this.id = null;
         this.scheduleDate = this.scheduleDate.plusDays(1);
         this.done = false;
     }
-    // TODO: ModelMapper에서 DTO로 넘겨주고 싶음.
-//    public ScheduleType getType(){
-//        if (goalCount > 0 | mon | tue | wed | thu | fri | sat | sun){
-//            return ScheduleType.ROUTINE;
-//        }else{
-//            return ScheduleType.TODO;
-//        }
-//    }
 
     // getRepeatDays() -> objectMapper.convertValue() 에서 컬럼으로 인식하고 변환해버림.. -.-
     public boolean[] calculateRepeatDays() {
@@ -146,13 +172,22 @@ public class Schedule extends BaseTime {
         return dayArray;
     }
 
-
-    public boolean getDone() {
-        return done;
+    public void verifyRoutine() {
+        if ((this.mon | this.tue | this.wed | this.thu | this.fri | this.sat | this.sun) && (
+            this.goalCount > 0)) {
+            throw new NotValidRoutineException("Repeat option of routine is not valid",
+                ErrorCode.ROUTINE_NOT_VALID);
+        }
     }
 
     public enum ScheduleType {
         TODO, ROUTINE
+    }
+
+    public void minusDoneCount() {
+        if (this.doneCount > 0) {
+            this.doneCount -= 1;
+        }
     }
 
     @Schema(description = "스케줄 생성 시 요청 클래스")
@@ -162,13 +197,13 @@ public class Schedule extends BaseTime {
     @NoArgsConstructor
     public static class ScheduleRequest {
 
-        @Schema(description = "사용자 DB ID", defaultValue = "1")
-        @NotNull
-        private Long userId;
-
         @Schema(description = "스케줄 이름", defaultValue = "술 마시기")
-        @NotNull
+        @NotBlank
         private String scheduleName;
+
+        @Schema(description = "잔소리 Id, 없는 경우 null")
+        @Nullable
+        private Long naggingId;
 
         @Schema(description = "n회 습관일 경우 목표 횟수", defaultValue = "0")
         private int goalCount;
@@ -177,12 +212,14 @@ public class Schedule extends BaseTime {
         private String scheduleTime;
 
         @Schema(description = "스케줄 수행 일자")
-        @NotNull
+        @NotBlank
         @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
+//        @JsonFormat(shape = Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
         private LocalDate scheduleDate;
 
-        @Schema(description = "스케줄 알람 시간")
-        @DateTimeFormat(pattern = "HH:mm:ss")
+        @Schema(description = "스케줄 알람 시간", defaultValue = "12:00:00")
+        @DateTimeFormat(iso = ISO.TIME)
+//        @JsonFormat(shape = Shape.STRING, pattern = "HH:mm:ss", timezone = "Asia/Seoul")
         private LocalTime alarmTime;
 
         @Schema(description = "월요일 반복 여부", defaultValue = "false")
@@ -205,9 +242,6 @@ public class Schedule extends BaseTime {
 
         @Schema(description = "일요일 반복 여부", defaultValue = "false")
         private boolean sun;
-
-        @Schema(description = "스케줄 유형(할일/습관)", defaultValue = "TODO")
-        private ScheduleType scheduleType;
     }
 
     @Schema(description = "단일 스케줄 조회 시 응답 클래스")
@@ -219,21 +253,21 @@ public class Schedule extends BaseTime {
 
         @Schema(description = "스케줄 ID", defaultValue = "2")
         private Long id;
-        @Schema(description = "사용자 DB ID", defaultValue = "1")
-        private Long userId;
+        @Schema(description = "잔소리 Id", defaultValue = "1")
+        private Long naggingId;
         @Schema(description = "n회 습관의 수행 목표 수", defaultValue = "0")
         private int goalCount;
-        @Schema(description = "n회 습관의 수행 완료 수", defaultValue = "0")
-        private int doneCount;
+        //        @Schema(description = "n회 습관의 수행 완료 수", defaultValue = "0")
+//        private int doneCount;
         @Schema(description = "스케줄 이름", defaultValue = "술 마시기")
         private String scheduleName;
         @Schema(description = "스케줄 수행 시간", defaultValue = "아무때나")
         private String scheduleTime;
-        @Schema(description = "스케줄 수행 일자")
+        @Schema(description = "스케줄 수행 일자", defaultValue = "2022-05-08")
         @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
         private LocalDate scheduleDate;
-        @Schema(description = "스케줄 알람 시간")
-        @DateTimeFormat(pattern = "HH:mm:ss")
+        @Schema(description = "스케줄 알람 시간", defaultValue = "12:00:00")
+        @DateTimeFormat(iso = ISO.TIME)
         private LocalTime alarmTime;
         @Schema(description = "스케줄 수행 여부", defaultValue = "false")
         private boolean done;
@@ -262,17 +296,26 @@ public class Schedule extends BaseTime {
     @NoArgsConstructor
     public static class ScheduleListResponse {
 
-        @Schema(description = "스케줄 ID", defaultValue = "2")
+        @Schema(description = "스케줄 ID", defaultValue = "22")
         private Long id;
-        @Schema(description = "스케줄 정렬 순서", defaultValue = "0")
-        private int seqNumber;
+        @Schema(description = "스케줄 원본 ID", defaultValue = "1")
+        private Long originalId;
+        @Schema(description = "n회 습관의 수행 목표 수", defaultValue = "0")
+        private int goalCount;
+        //        @Schema(description = "n회 습관의 수행 완료 수", defaultValue = "0")
+//        private int doneCount;
+//        @Schema(description = "잔소리 Id", defaultValue = "1")
+//        private Long naggingId;
         @Schema(description = "스케줄 이름", defaultValue = "술 마시기")
         private String scheduleName;
         @Schema(description = "스케줄 수행 시간", defaultValue = "아무때나")
         private String scheduleTime;
-        @Schema(description = "수행 완료 여부", defaultValue = "false", allowableValues = {"true",
-            "false"})
-        private boolean isDone;
+        @Schema(description = "수행 완료 여부\n\n"
+            + "true : 수행 완료\n\n"
+            + "false : 미수행\n\n"
+            + "null : 미룸", defaultValue = "false", allowableValues = {"true",
+            "false", "null"})
+        private boolean done;
         @Schema(description = "스케줄 유형(할일/습관)", defaultValue = "todo")
         private ScheduleType scheduleType;
 
@@ -289,5 +332,30 @@ public class Schedule extends BaseTime {
         private Long id;
         @Schema(description = "스케줄 이름", defaultValue = "술 마시기")
         private String scheduleName;
+    }
+
+    @Schema(description = "스케줄 순서 변경 시 요청 클래스")
+    @Getter
+    @Setter
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public static class ArrayListRequest {
+
+        @Schema(description = "순서를 스위칭할 스케줄의 original ID")
+        private Long oneOriginalId;
+
+        @Schema(description = "순서를 스위칭할 또 다른 스케줄의 original ID")
+        private Long theOtherOriginalId;
+    }
+
+    @Getter
+    @Setter
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public static class SchedulePush {
+
+        private String title;
+        private String targetToken;
+        private String body;
     }
 }
