@@ -1,16 +1,39 @@
 package com.jasik.momsnaggingapi.domain.user;
 
-import com.jasik.momsnaggingapi.domain.common.BaseTime;
+import com.jasik.momsnaggingapi.infra.common.BaseTime;
+
+import io.jsonwebtoken.Claims;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
+
+import org.hibernate.annotations.ColumnDefault;
+import org.hibernate.annotations.DynamicUpdate;
 import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.security.core.GrantedAuthority;
 
 import javax.persistence.*;
+import javax.security.auth.Subject;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.Collection;
+import com.jasik.momsnaggingapi.infra.common.StringListConverter;
+import io.swagger.v3.oas.annotations.media.Schema;
+import java.util.List;
+import javax.persistence.Column;
+import javax.persistence.Convert;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
 
 @Entity
-@Getter
+@Getter @Setter
 @NoArgsConstructor
 public class User extends BaseTime {
 
@@ -18,7 +41,7 @@ public class User extends BaseTime {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private int naggingLevel;
+    private Integer naggingLevel;
     private String nickName;
     private String email;
     private String provider;
@@ -26,33 +49,61 @@ public class User extends BaseTime {
     private String personalId;
     private String device;
     private String profileImage;
+    private String firebaseToken;
 
     @Column(columnDefinition = "varchar(30) default 'MEMBER'")
-    private String role;
-    @Column(columnDefinition = "TEXT")
-    private String statusMsg;
+    private String role = "MEMBER";
+    @Column(columnDefinition = "TEXT default '오늘 하루도 파이팅 🔥")
+    private String statusMsg = "";
 
     @Column(columnDefinition = "boolean default true")
-    private boolean allowGeneralNotice;
+    private Boolean allowTodoNotice;
     @Column(columnDefinition = "boolean default true")
-    private boolean allowTodoNotice;
+    private Boolean allowRoutineNotice;
     @Column(columnDefinition = "boolean default true")
-    private boolean allowRoutineNotice;
+    private Boolean allowWeeklyNotice;
     @Column(columnDefinition = "boolean default true")
-    private boolean allowWeeklyNotice;
-    @Column(columnDefinition = "boolean default true")
-    private boolean allowOtherNotice;
+    private Boolean allowOtherNotice;
+
+    @Convert(converter = StringListConverter.class)
+    @Column(columnDefinition = "json")
+    private List<String> routineOrder;
+
+    public void updateRoutineOrder(List<String> newOrder) {
+        this.routineOrder = newOrder;
+    }
 
     @Builder
-    public User(int naggingLevel, String nickName, String email,
-                String provider, String providerCode, String personalId, String device) {
-        this.naggingLevel = naggingLevel;
+    public User(String nickName, String email,
+        String provider, String providerCode, String personalId, String device, String firebaseToken) {
         this.nickName = nickName;
         this.email = email;
         this.provider = provider;
         this.providerCode = providerCode;
         this.personalId = personalId;
         this.device = device;
+        this.firebaseToken = firebaseToken;
+        this.naggingLevel = 0;
+        this.allowTodoNotice = true;
+        this.allowRoutineNotice = true;
+        this.allowWeeklyNotice = true;
+        this.allowOtherNotice = true;
+        this.statusMsg = "오늘 하루도 파이팅 🔥";
+    }
+
+    public User(Claims claims) {
+        this.id = Long.valueOf(claims.getSubject());
+        this.personalId = claims.get("id").toString();
+//        this.email = claims.get("email").toString();
+        this.provider = claims.get("provider").toString();
+    }
+
+    @Schema(description = "사용자 관련 기본 응답 클래스")
+    @Getter @Setter
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public static class Response {
+        private Long id;
     }
 
     @Schema(description = "사용자 조회 시 응답 클래스")
@@ -65,13 +116,13 @@ public class User extends BaseTime {
         private String provider;
         private String nickName;
         private String personalId;
-        private int naggingLevel;
+        private Integer naggingLevel;
         private String device;
-        private boolean allowGeneralNotice;
-        private boolean allowTodoNotice;
-        private boolean allowRoutineNotice;
-        private boolean allowWeeklyNotice;
-        private boolean allowOtherNotice;
+        private Boolean allowTodoNotice;
+        private Boolean allowRoutineNotice;
+        private Boolean allowWeeklyNotice;
+        private Boolean allowOtherNotice;
+        private String statusMsg;
     }
 
     @Schema(description = "로그인 요청 클래스")
@@ -83,6 +134,10 @@ public class User extends BaseTime {
         private String provider;
         @Schema(description = "플랫폼 인증 코드")
         private String code;
+        @Schema(description = "디바이스(IOS | AOS)")
+        private String device;
+        @Schema(description = "파이어베이스 엑세스 토큰")
+        private String firebaseToken;
     }
 
     @Schema(description = "로그인/회원가입 응답 클래스")
@@ -99,31 +154,33 @@ public class User extends BaseTime {
     @AllArgsConstructor
     @NoArgsConstructor
     public static class CreateRequest {
-        @Schema(description = "소셜로그인 플랫폼")
+        @Schema(description = "소셜로그인 플랫폼(Kakao | Google)")
         private String provider;
         @Schema(description = "사용자 이메일")
         private String email;
         @Schema(description = "소셜 코드")
         private String code;
-        @Schema(description = "디바이스")
+        @Schema(description = "디바이스(IOS | AOS)")
         private String device;
         @Schema(description = "아이디")
         private String personalId;
         @Schema(description = "호칭")
         private String nickname;
+        @Schema(description = "파이어베이스 엑세스 토큰")
+        private String firebaseToken;
     }
 
-    @Getter
+    @Getter @Setter
     @AllArgsConstructor
     @NoArgsConstructor
     public static class UpdateRequest {
         private String nickName;
-        private int naggingLevel;
-        private boolean allowGeneralNotice;
-        private boolean allowTodoNotice;
-        private boolean allowRoutineNotice;
-        private boolean allowWeeklyNotice;
-        private boolean allowOtherNotice;
+        private Integer naggingLevel;
+        private Boolean allowTodoNotice;
+        private Boolean allowRoutineNotice;
+        private Boolean allowWeeklyNotice;
+        private Boolean allowOtherNotice;
+        private String statusMsg;
     }
 
     @Schema(description = "아이디 중복확인 응답 클래스")
